Make eos extraction default
