+++ dalm/eval/eval_retriever_only.py
@@ -93,11 +93,6 @@ def parse_args() -> Namespace:
         action="store_true",
         help="Whether the model is autoregressive or not",
     )
-    parser.add_argument(
-        "--extract_eos_only",
-        action="store_true",
-        help="Whether to extract only the last token of the sequence",
-    )
     args = parser.parse_args()
 
     return args
@@ -116,7 +111,6 @@ def evaluate_retriever(
     torch_dtype: Literal["float16", "bfloat16"] = "float16",
     top_k: int = 10,
     is_autoregressive: bool = False,
-    extract_eos_only: bool = False,
 ) -> None:
     """Runs rag evaluation. See `dalm eval-retriever --help for details on params"""
     test_dataset = load_dataset(dataset_or_path)
@@ -126,7 +120,6 @@ def evaluate_retriever(
         get_peft=False,
         use_bnb=False,
         is_autoregressive=is_autoregressive,
-        extract_eos_only=extract_eos_only,
     )
     retriever_tokenizer = retriever_model.tokenizer
 
@@ -194,7 +187,6 @@ def main() -> None:
         torch_dtype=args.torch_dtype,
         top_k=args.top_k,
         is_autoregressive=args.is_autoregressive,
-        extract_eos_only=args.extract_eos_only,
     )
