+++ dalm/models/retriever_only_base_model.py
@@ -1,25 +1,10 @@
-import warnings
 from typing import List, Optional, Union
 
 import torch
 from peft import LoraConfig, PeftModel, TaskType, get_peft_model
 from transformers import AutoModel, AutoModelForCausalLM, AutoTokenizer, BitsAndBytesConfig
 
-
-def eos_mask(mask: torch.Tensor, padding: str = "left") -> torch.Tensor:
-    """
-    Returns a mask that only masks everything else but the last token of each sequence.
-    """
-    new_mask = torch.zeros_like(mask)
-
-    if padding == "right":
-        ones_counts = mask.sum(dim=1)
-        indices = (torch.arange(mask.size(0)), ones_counts - 1)
-        new_mask[indices] = 1
-    else:
-        new_mask[:, -1] = 1
-
-    return new_mask
+from dalm.utils import eos_mask
 
 
 class AutoModelForSentenceEmbedding(torch.nn.Module):
@@ -30,7 +15,6 @@ class AutoModelForSentenceEmbedding(torch.nn.Module):
         use_bnb: bool = True,
         get_peft: bool = True,
         is_autoregressive: bool = False,
-        extract_eos_only: bool = False,
     ) -> None:
         super(AutoModelForSentenceEmbedding, self).__init__()
 
@@ -56,17 +40,19 @@ class AutoModelForSentenceEmbedding(torch.nn.Module):
             self.tokenizer.add_eos_token = True
             self.tokenizer.pad_token = self.tokenizer.eos_token
 
-        self.extract_eos_only = extract_eos_only
-
     def forward(self, input_ids: torch.Tensor, attention_mask: torch.Tensor) -> torch.Tensor:
+        # Note: previously there existed two ways of extracting embedding from autoregressive models
+        # 1. average over the last hidden states of all tokens
+        # 2. take the last hidden state of the EOS token
+        #
+        # 1 was outperformed by 2, so now 2 is only used
         if self.is_autoregressive:
             # we take the last hidden state of the model
             token_embeddings = self.model(
                 input_ids, attention_mask=attention_mask, output_hidden_states=True, return_dict=True
             ).hidden_states[-1]
 
-            if self.extract_eos_only:
-                attention_mask = eos_mask(attention_mask)
+            attention_mask = eos_mask(attention_mask)
         else:
             # First element of model_output contains all token embeddings
             token_embeddings = self.model(input_ids, attention_mask)[0]
@@ -122,7 +108,3 @@ class AutoModelForSentenceEmbedding(torch.nn.Module):
             target_modules=target_modules
             or (["key", "query", "value"] if not is_autoregressive else ["q_proj", "v_proj"]),
         )
-
-    def __post_init__(self) -> None:
-        if not self.is_autoregressive and self.extract_eos_only:
-            warnings.warn("eos_extraction_method is ignored for non-autoregressive models", stacklevel=2)
