+++ dalm/training/retriever_only/train_retriever_only.py
@@ -168,11 +168,6 @@ def parse_args() -> Namespace:
         action="store_true",
         help="Whether model is an auto-regressive model/clm ",
     )
-    parser.add_argument(
-        "--extract_eos_only",
-        action="store_true",
-        help="Whether to extract only the last token of the sequence",
-    )
     args = parser.parse_args()
 
     return args
@@ -206,7 +201,6 @@ def train_retriever(
     use_peft: bool = True,
     use_bnb: bool = True,
     is_autoregressive: bool = False,
-    extract_eos_only: bool = False,
 ) -> None:
     # Get the passed in vars before beginning training, in case we report training
     args = dict(locals())
@@ -244,7 +238,6 @@ def train_retriever(
         use_bnb=use_bnb,
         get_peft=use_peft,
         is_autoregressive=is_autoregressive,
-        extract_eos_only=extract_eos_only,
     )
     tokenizer = model.tokenizer
 
@@ -463,7 +456,6 @@ def main() -> None:
         use_peft=args.use_peft,
         use_bnb=args.use_bnb,
         is_autoregressive=args.is_autoregressive,
-        extract_eos_only=args.extract_eos_only,
     )
