+++ dalm/utils.py
@@ -1,6 +1,7 @@
 import os
 
 import datasets
+import torch
 from datasets import Dataset
 
 
@@ -16,3 +17,19 @@ def load_dataset(dataset_or_path: str | Dataset) -> Dataset:
         if os.path.isdir(dataset_or_path)
         else datasets.load_dataset("csv", data_files=dataset_or_path)["train"]
     )
+
+
+def eos_mask(mask: torch.Tensor, padding: str = "left") -> torch.Tensor:
+    """
+    Returns a mask that only masks everything else but the last token of each sequence.
+    """
+    new_mask = torch.zeros_like(mask)
+
+    if padding == "right":
+        ones_counts = mask.sum(dim=1)
+        indices = (torch.arange(mask.size(0)), ones_counts - 1)
+        new_mask[indices] = 1
+    else:
+        new_mask[:, -1] = 1
+
+    return new_mask
